Commit final?

+++ Casino/src/aplicacion/Producto.java
@@ -7,10 +7,10 @@ package aplicacion;
  */
 public class Producto {
     private String nombre;
-    private int existencias;
+    private float existencias;
     private float precio_compra;
 
-    public Producto(String nombre, int existencias, float precio_compra) {
+    public Producto(String nombre, float existencias, float precio_compra) {
         this.nombre = nombre;
         this.existencias = existencias;
         this.precio_compra = precio_compra;
@@ -24,7 +24,7 @@ public class Producto {
         this.nombre = nombre;
     }
 
-    public int getExistencias() {
+    public float getExistencias() {
         return existencias;
     }
